add unit test

+++ Django/orders/models.py
@@ -4,6 +4,6 @@ class Order(models.Model):
     marketplace = models.fields.CharField(max_length=100)
     ref_id = models.fields.CharField(max_length=100, primary_key=True, editable=False)
     date = models.fields.DateField(blank=True, null=True)
-    amount = models.fields.DecimalField(max_digits=10, decimal_places=2)
+    amount = models.fields.FloatField()
     currency = models.fields.CharField(max_length=3)
